requirements.txt
python-dotenv>=1.0.1
numpy>=1.26.0
pypdf>=4.2.0
cryptography>=42.0.0
python-docx>=1.1.2
streamlit>=1.39.0
google-generativeai>=0.8.3
